docs(README.md): add instructions for starting environment with nvidia-container-toolkit to help users with dependency issues fix(openai_types.py): change 'created' field type from int to float in ChatCompletion model to accurately represent time values

+++ functionary/openai_types.py
@@ -70,5 +70,5 @@ class Choice(BaseModel):
 class ChatCompletion(BaseModel):
     id: str
     object: str = "chat.completion"
-    created: int = Field(default_factory=time.time)
+    created: float = Field(default_factory=time.time)
     choices: List[Choice]
